refactor: reorganize project role and application structure
- Updated import paths for ProjectRole and ProjectRoleApplication to reflect new directory structure under bounded contexts.
- Introduced new entities, repositories, and use cases for project roles and applications, enhancing modularity and maintainability.
- Added controllers and DTOs for handling project role applications, improving API functionality.
- Implemented command handlers for creating and managing project role applications, ensuring better encapsulation of business logic.
