+++ apps/server/src/contexts/contexts.module.ts
@@ -3,10 +3,10 @@ import { UserInfrastructure } from './user/infrastructure/user.infrastructure';
 import { ProjectInfrastructure } from './project/infrastructure/project.infrastructure';
 import { ProfileInfrastructure } from './profile/infrastructure/profile.infrastructure';
 import { TechStackInfrastructure } from './techstack/infrastructure/techstack.infrastructure';
-import { ProjectRoleInfrastructure } from './project-role/infrastructure/project-role.infrastructure';
+import { ProjectRoleInfrastructure } from './project/bounded-contexts/project-role/infrastructure/project-role.infrastructure';
 import { GithubInfrastructure } from './github/infrastructure/github.infrastructure';
 import { CategoryInfrastructure } from './category/infrastructure/category.infrastructure';
-import { ProjectRoleApplicationInfrastructure } from './project-role-application/infrastructure/project-role-application.infrastructure';
+import { ProjectRoleApplicationInfrastructure } from './project/bounded-contexts/project-role-application/infrastructure/project-role-application.infrastructure';
 import { MailingInfrastructure } from '@/mailing/infrastructure/mailing.infrastructure';
 
 @Module({
