+++ apps/server/src/contexts/project/bounded-contexts/project-goals/domain/project-goals.entity.ts
