+++ apps/server/src/contexts/project/bounded-contexts/project-role-application/infrastructure/project-role-application.infrastructure.ts
@@ -1,16 +1,17 @@
 import { Module } from '@nestjs/common';
-import { PROJECT_ROLE_APPLICATION_REPOSITORY_PORT } from '@/contexts/project-role-application/use-cases/ports/project-role-application.repository.port';
+import { PROJECT_ROLE_APPLICATION_REPOSITORY_PORT } from '@/contexts/project/bounded-contexts/project-role-application/use-cases/ports/project-role-application.repository.port';
 import { PersistenceInfrastructure } from '@/persistence/persistence.infrastructure';
-import { PrismaProjectRoleApplicationRepository } from '@/contexts/project-role-application/infrastructure/repositories/prisma.project-role-application.repository';
+import { PrismaProjectRoleApplicationRepository } from '@/contexts/project/bounded-contexts/project-role-application/infrastructure/repositories/prisma.project-role-application.repository';
 import { projectRoleApplicationUseCases } from '../use-cases/project-role-application.use-cases';
-import { PROJECT_ROLE_REPOSITORY_PORT } from '@/contexts/project-role/use-cases/ports/project-role.repository.port';
-import { PrismaProjectRoleRepository } from '@/contexts/project-role/infrastructure/repositories/prisma.project-role.repository';
+import { PROJECT_ROLE_REPOSITORY_PORT } from '@/contexts/project/bounded-contexts/project-role/use-cases/ports/project-role.repository.port';
+import { PrismaProjectRoleRepository } from '@/contexts/project/bounded-contexts/project-role/infrastructure/repositories/prisma.project-role.repository';
 import { PROJECT_REPOSITORY_PORT } from '@/contexts/project/use-cases/ports/project.repository.port';
 import { PrismaProjectRepository } from '@/contexts/project/infrastructure/repositories/prisma.project.repository';
 import { MAILING_SERVICE_PORT } from '@/mailing/ports/mailing.service.port';
 import { ResendMailingService } from '@/mailing/infrastructure/resend.mailing.service';
 import { USER_REPOSITORY_PORT } from '@/contexts/user/use-cases/ports/user.repository.port';
 import { PrismaUserRepository } from '@/contexts/user/infrastructure/repositories/prisma.user.repository';
+import { ProjectRoleApplicationController } from './controllers/project-role-application.controller';
 
 @Module({
   imports: [PersistenceInfrastructure],
@@ -37,6 +38,7 @@ import { PrismaUserRepository } from '@/contexts/user/infrastructure/repositorie
       useClass: PrismaUserRepository,
     },
   ],
+  controllers: [ProjectRoleApplicationController],
   exports: [
     PROJECT_ROLE_APPLICATION_REPOSITORY_PORT,
     PROJECT_ROLE_REPOSITORY_PORT,
