+++ apps/server/src/contexts/project/bounded-contexts/project-role-application/infrastructure/repositories/prisma.project-role-application.mapper.ts
