+++ apps/server/src/contexts/project/bounded-contexts/project-role-application/use-cases/queries/get-all-project-application.query.ts
