+++ apps/server/src/contexts/project/bounded-contexts/project-role-application/use-cases/queries/project-role-application.queries.ts
