+++ apps/server/src/contexts/project/bounded-contexts/project-role/domain/project-role.entity.ts
