+++ apps/server/src/contexts/project/bounded-contexts/project-role/use-cases/queries/get-project-role-by-id.query.ts
