+++ apps/server/src/contexts/project/infrastructure/project.infrastructure.ts
@@ -2,9 +2,9 @@ import { Module } from '@nestjs/common';
 import { PROJECT_REPOSITORY_PORT } from '@/contexts/project/use-cases/ports/project.repository.port';
 import { PersistenceInfrastructure } from '@/persistence/persistence.infrastructure';
 import { projectUseCases } from '@/contexts/project/use-cases/project.use-cases';
-import { PROJECT_ROLE_REPOSITORY_PORT } from '@/contexts/project-role/use-cases/ports/project-role.repository.port';
+import { PROJECT_ROLE_REPOSITORY_PORT } from '@/contexts/project/bounded-contexts/project-role/use-cases/ports/project-role.repository.port';
 import { PrismaProjectRepository } from '@/contexts/project/infrastructure/repositories/prisma.project.repository';
-import { PrismaProjectRoleRepository } from '@/contexts/project-role/infrastructure/repositories/prisma.project-role.repository';
+import { PrismaProjectRoleRepository } from '@/contexts/project/bounded-contexts/project-role/infrastructure/repositories/prisma.project-role.repository';
 import { TECHSTACK_REPOSITORY_PORT } from '@/contexts/techstack/use-cases/ports/techstack.repository.port';
 import { PrismaTechStackRepository } from '@/contexts/techstack/infrastructure/repositories/prisma.techstack.repository';
 import { ProjectController } from '@/contexts/project/infrastructure/controllers/project.controller';
@@ -19,7 +19,7 @@ import { ENCRYPTION_SERVICE_PORT } from '@/contexts/encryption/ports/encryption.
 import { EncryptionService } from '@/contexts/encryption/infrastructure/encryption.service';
 import { InMemoryProjectRepository } from './repositories/mock.project.repository';
 import { InMemoryTechStackRepository } from '@/contexts/techstack/infrastructure/repositories/mock.techstack.repository';
-import { InMemoryProjectRoleRepository } from '@/contexts/project-role/infrastructure/repositories/mock.project-role.repository';
+import { InMemoryProjectRoleRepository } from '@/contexts/project/bounded-contexts/project-role/infrastructure/repositories/mock.project-role.repository';
 import { CATEGORY_REPOSITORY_PORT } from '@/contexts/category/use-cases/ports/category.repository.port';
 import { PrismaCategoryRepository } from '@/contexts/category/infrastructure/repositories/prisma.category.repository';
 import { PROFILE_REPOSITORY_PORT } from '@/contexts/profile/use-cases/ports/profile.repository.port';
